# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BotMarket API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotType {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  emoji: string;
  features: string[];
  popular: boolean;
  /** @nullable */
  imageUrl?: string | null;
  serverMonthlyPrice?: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  botTypeId: number;
  botTypeName: string;
  clientName: string;
  clientPhone: string;
  telegramId: string;
  /** @nullable */
  telegramUsername?: string | null;
  status: OrderStatus;
  totalPrice: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  referralCode?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderInput {
  botTypeId: number;
  clientName: string;
  clientPhone: string;
  telegramId: string;
  /** @nullable */
  telegramUsername?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  referralCode?: string | null;
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  pending: "pending",
  confirmed: "confirmed",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface DashboardStats {
  totalOrders: number;
  pendingOrders: number;
  completedOrders: number;
  totalRevenue: number;
  totalUsers: number;
  popularBot: string;
  recentOrders?: Order[];
}

export interface ReferralInfo {
  code: string;
  referrerName: string;
  totalReferrals: number;
}