# lib/api-zod/src/generated/types/botType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BotMarket API
 * OpenAPI spec version: 0.1.0
 */

export interface BotType {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  emoji: string;
  features: string[];
  popular: boolean;
  /** @nullable */
  imageUrl?: string | null;
  serverMonthlyPrice?: number;
}