# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BotMarket API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from "./order";

export interface DashboardStats {
  totalOrders: number;
  pendingOrders: number;
  completedOrders: number;
  totalRevenue: number;
  totalUsers: number;
  popularBot: string;
  recentOrders?: Order[];
}