# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BotMarket API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  botTypeId: number;
  botTypeName: string;
  clientName: string;
  clientPhone: string;
  telegramId: string;
  /** @nullable */
  telegramUsername?: string | null;
  status: OrderStatus;
  totalPrice: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  referralCode?: string | null;
  createdAt: string;
  updatedAt?: string;
}