# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BotMarket API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  botTypeId: number;
  clientName: string;
  clientPhone: string;
  telegramId: string;
  /** @nullable */
  telegramUsername?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  referralCode?: string | null;
}